Core - confusing variable name left SMS confirmation broken
In the transition to python-messaging 0.5.9 the SMS PDU encoding no
longer takes an argument for request of confirmation, relying on the
caller setting a variable within the instantiated message. Unfortuately
the core uses a varable called 'status_request', whilst the objects
internal one is 'request_status'. This patch modifies the core to set
the correct value again.