Core - Reorder property setting logic
The existing logic was flawed in that it failed to set the Device property
when the plugin was reused after hotplug. This only created a problem when
a device was reinserted but the kernel assigned a different portname for the
dataport, such a case is when a stick is pulled out whilst a ppp connection
is in progress.