Core - Ensure formatting applies to correct get_network_info command
When several clients are querying the core for status information,
it is possible, and even likely if signal driven, that multiple
requests for get_network_info() get processed at the same time. This
leads to the situation where commands are interleaved between the
+COPS format definition and the request for network info, resulting
in the information being returned in the wrong format. Making the
format specifier part of the request for info alleviates this.