Core - Change Enable / Authentication semantics to be like Modem Manager
We used to assume that Authentication would always be followed by SIM
initialisation but we cannot do that now as Network Manager's applet does
authentication as part of spotting there's a new device just been plugged
in. This means that we could have have the SendPin() called on its own,
or as part of a sequence like Enable() / SimpleConnect(). Change Sendpin()
to save a timestamp when it was run, and Enable() to check if DELAY has
elapsed before starting the SIM initialisation.