Tests - remove network provider import
Importing by filename no longer functions since python 2.6 so, we
can't rely on the developer directory structure to import the
current network database. But equally we don't want to test an older
version of the networks.py file that's already installed on the
system. We test this more fully later with a relative path import.