Disable OSX support until we can find out about the C module
The OSX serial ports are underpinned by a C module, but there's no
indication of its licence or provenance. We can't ship the Debian
packages with it included so remove the C module and disable the OS
plugin until we can resolve by clarification or rewrite. As ever,
patches are most welcome.