Udev rule update - be more like MM, avoids usb_modeswitch bug
The usb_modeswitch rules call the /lib/udev/usb_modeswitch script
with an argument constructed out of '%b/%k'. Unfortunately when %b
is populated usb_modeswitch no longer finds the device, so switching
is not performed. Wader was using the udev rules language in such a
way that %b was populated for example:
Modem Manager:
	/lib/udev/usb_modeswitch '/1.3:1.0'
Wader:
	/lib/udev/usb_modeswitch '1.3:1.0/1.3:1.0'

The easy fix for us is to rewrite our udev rules to be more like
Modem Manager's, but potentially the same problem could be caused by
an unrelated udev rule.