Expand get_usage_for_day provider test
Previously the get_usage_for_day test would be run with a base time
that was dependent upon the time of day the test was being run. This
caused intermittant failures depending upon if the start and end
times of the items straddled the date line. This patch splits the
test into two and runs each with a fixed date/time for a predictable
outcome.
Note: after applying this patch I expect one test to fail until a
following patch fixes the problem.